% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dtq.R
\name{append.log}
\alias{append.log}
\title{Prepare and call append log to cache}
\usage{
append.log(te, dtcall, pt)
}
\arguments{
\item{te}{environment}

\item{dtcall}{call}

\item{pt}{numeric}
}
\description{
Prepare and call append log to cache
}

