% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dtq.R
\name{deparse.dtq.call}
\alias{deparse.dtq.call}
\title{deparse data.table query call}
\usage{
deparse.dtq.call(x)
}
\arguments{
\item{x}{list of calls}
}
\value{
list of character
}
\description{
deparse data.table query call
}

