% Generated by roxygen2: do not edit by hand
\name{deparse_and_paste_arg}
\alias{deparse_and_paste_arg}
\title{deparse argument names and their values}
\usage{
deparse_and_paste_arg(args.names, args)
}
\arguments{
\item{args.names}{character}

\item{args}{language objects}
}
\value{
character
}
\description{
deparse argument names and their values
}

