% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dtq.R
\name{deparse_char}
\alias{deparse_char}
\title{deparse object to character}
\usage{
deparse_char(x)
}
\arguments{
\item{x}{object to deparse}
}
\value{
character scalar
}
\description{
deparse object to character
}

