% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dtq.R
\name{depth.dtq.call}
\alias{depth.dtq.call}
\title{dtq depth}
\usage{
depth.dtq.call(x, apply.depth = getOption("dtq.apply.depth", 20L))
}
\arguments{
\item{x}{list of calls}

\item{apply.depth}{integer limit of recursive call in chain}
}
\value{
depth of tested queries in chain, points to the `src` object.
}
\description{
dtq depth
}
\note{
Currently not used, process in using `dtq` class object `depth` method because dtq class store calls casted to recursive lists of calls. Calls are matched to `[.data.table` and that makes `depth` function catch more cases.
}

