% Generated by roxygen2: do not edit by hand
\name{do.dtq.log}
\alias{do.dtq.log}
\title{Check if proceed with dtq logging}
\usage{
do.dtq.log(te)
}
\arguments{
\item{te}{environment}
}
\description{
Check if proceed with dtq logging
}
\details{
Uses \code{dtq.log.exclude} and \code{dtq.log.include} options to define if logging should be done based on the \code{te} environment.
}

