% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dtq.R
\docType{package}
\name{dtq-package}
\alias{dtq-package}
\title{dtq-package}
\description{
After loading package by default it start to log data.table queries and their processing details, logs can be accessed by \link{dtl} function.
}
\author{
Jan Gorecki
}

