% Generated by roxygen2: do not edit by hand
\docType{class}
\name{dtq.log}
\alias{.DTQ}
\alias{dtq.log}
\title{data.table queries log storage}
\format{An R6 class object.}
\usage{
dtq.log
}
\description{
data.table queries log storage
}
\details{
Environment to store data.table queries, use \link{dtl} to access formatted logs.
}
\seealso{
\link{dtl}, \link{dtq}
}
\keyword{datasets}

