% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dtq_cms_to_cfs}
\alias{dtq_cms_to_cfs}
\alias{dtq_cms_to_kcfs}
\alias{dtq_kcfs_to_cms}
\alias{dtq_cfs_to_cms}
\title{Cubic Metres per Second to Cubic Feet per Second}
\usage{
dtq_cms_to_cfs(x)

dtq_cms_to_kcfs(x)

dtq_kcfs_to_cms(x)

dtq_cfs_to_cms(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector.
}
\description{
Cubic Metres per Second to Cubic Feet per Second
}
\section{Functions}{
\itemize{
\item \code{dtq_cms_to_kcfs}: cms to kcfs

\item \code{dtq_kcfs_to_cms}: kcfs to cms

\item \code{dtq_cfs_to_cms}: cfs to cms
}}

\examples{
dtq_cms_to_cfs(1)
dtq_cms_to_kcfs(1)
dtq_cfs_to_cms(1)
dtq_kcfs_to_cms(1)
}
