% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge-to-stage.R
\name{dtq_discharge_to_stage}
\alias{dtq_discharge_to_stage}
\title{Convert Discharge to Stage}
\usage{
dtq_discharge_to_stage(x, r1d, dtt = "DateTime", colname = "Discharge",
  stage = "Stage", n = 0L, rate_down = Inf, rate_up = rate_down,
  units = dtt_units(x[[dtt]]))
}
\arguments{
\item{x}{A dts data frame}

\item{r1d}{A data frame of the discharge and stage values for a transect 
from a 1-dimensional model.}

\item{dtt}{A string specifying the column with the Date or POSIXct values.}

\item{colname}{A character vector specifying the column(s) with the discharge and/or stage data.}

\item{stage}{A string of the name of the column with the stage values.}

\item{n}{A count of the number of time units until the stage responds.}

\item{rate_down}{A positive number of the maximum absolute rate when dropping the discharge.}

\item{rate_up}{A positive number of the maximum rate when increasing the discharge.}

\item{units}{A string of the units.}
}
\value{
The updated dtq data frame with a stage column.
}
\description{
Convert Discharge to Stage
}
