% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dtq.R
\name{dtt}
\alias{dtt}
\title{Get data.table queries logs}
\usage{
dtt(chain.only = FALSE, purge = FALSE)
}
\arguments{
\item{chain.only}{logical if \emph{TRUE} then aggregated logs to full chains}

\item{purge}{logical when true will clear cache and return empty data.table}
}
\description{
Get data.table queries logs
}

