% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dtq.R
\name{is.dtq.call}
\alias{is.dtq.call}
\title{is.dtq.call}
\usage{
is.dtq.call(x)
}
\arguments{
\item{x}{call or list of calls}
}
\value{
logical scalar or list of logical scalars
}
\description{
is.dtq.call
}

