% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dtq.R
\name{match.dtq.call}
\alias{match.dtq.call}
\title{match.dtq.call}
\usage{
match.dtq.call(x)
}
\arguments{
\item{x}{call or list of calls, valid data.table query call}
}
\value{
the same call(s) but with matched arguments
}
\description{
match.dtq.call
}
\note{
currently not even required during processing as dtq catches already matched call, but can be utilized in manual catching dtq
}
\seealso{
\link{is.dtq.call}
}

