% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dtq.R
\name{dtl}
\alias{dtl}
\title{Get data.table queries logs}
\usage{
dtl(print = FALSE, chain = FALSE, purge = FALSE)
}
\arguments{
\item{print}{logical if \emph{TRUE} then will remove \emph{list} type columns from the result.}

\item{chain}{logical if \emph{TRUE} then aggregated logs to full chains}

\item{purge}{logical when \emph{TRUE} it will clear dtq logs and return empty data.table}
}
\description{
Get data.table queries logs
}
\details{
\emph{timestamp} represents the time of query (or chain) log after it is evaluated.
}

