% Generated by roxygen2: do not edit by hand
\docType{package}
\name{dtq-package}
\alias{dtq-package}
\title{dtq-package}
\description{
After loading package by default it start to log data.table queries and their processing details, logs can be accessed by \link{dtl} function.
}
\author{
Jan Gorecki
}

